OmniInsight/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.1.0
shap>=0.44.0
streamlit>=1.30.0
openai>=1.0.0
PyYAML>=6.0.0
